requirements.txt
# requirements.txt
python-dotenv==1.0.0
fire==0.5.0
torch==2.0.1
transformers==4.30.2